requirements.txt
python-telegram-bot==20.0
python-dotenv
gspread
oauth2client
apscheduler
